+++ lib/block.mli
@@ -94,8 +94,14 @@ val set_variables: t -> (string * string) list
 val unset_variables: t -> string list
 (** [unset_variable t] is the list of environment variable to unset *)
 
-val required_packages: t -> string list
-(** [required_packages t] is the names of the required packages *)
+val explicit_required_packages: t -> string list
+(** [explicit_required_packages t] returns the list of packages explicitly required by the user
+    through require-package labels in the block [t]. *)
+
+val required_libraries: t -> (Library.Set.t, string) Result.result
+(** [required_libraries t] returns the set of libaries that are loaded through [#require]
+    statements in the block [t]. Always returns an empty set if [t] isn't a toplevel
+    block. *)
 
 val skip: t -> bool
 (** [skip t] is true iff [skip] is in the labels of [t]. *)
@@ -130,3 +136,11 @@ val labels_of_string:
   string ->
   (string * ([`Eq | `Neq | `Gt | `Ge | `Lt | `Le] * string) option) list
 (** [labels_of_string s] cuts [s] into a list of labels. *)
+
+val require_from_line : string -> (Library.Set.t, string) Result.result
+(** [require_from_line line] returns the set of libraries imported by the
+    #require statement on [line] or an empty set if [line] is not a require
+    statement. *)
+
+val require_from_lines : string list -> (Library.Set.t, string) Result.result
+(** Same as [require_from_line] but aggregated over several lines *)