+++ lib/mdx.mli
@@ -26,9 +26,12 @@
 module Output = Output
 module Cram = Cram
 module Toplevel = Toplevel
+module Library = Library
 module Block = Block
 module Migrate_ast = Migrate_ast
 module Compat = Compat
+module Util = Util
+module Prelude = Prelude
 
 (** {2 Lines} *)